refactor: clean up line spec builder
