+++ package.json
@@ -25,6 +25,8 @@
 		"g": "plop component",
 		"generate": "plop component",
 		"lint": "eslint src --ext .ts,.tsx,.js,.jsx",
+		"ts": "yarn tsc",
+		"tsc": "tsc --noEmit --skipLibCheck",
 		"prepack": "yarn build",
 		"pack-test": "rm -rf dist & npm pack",
 		"start": "yarn storybook",
