+++ src/constants.ts
@@ -51,3 +51,6 @@ export const TRELLIS_PADDING = 0.2;
 
 // ratio that each opacity is divded by when hovering or highlighting from legend
 export const HIGHLIGHT_CONTRAST_RATIO = 5;
+
+// signal names
+export const BACKGROUND_COLOR = 'backgroundColor';
