+++ src/types/index.ts
@@ -13,3 +13,4 @@
 export * from './Chart';
 export * from './specBuilderTypes';
 export * from './SpectrumVizColors';
+export * from './supplementalVegaTypes';
