+++ src/types/supplementalVegaTypes.ts
@@ -1,0 +1,3 @@
+export type ProductionRuleTests<T> = ({
+	test?: string;
+} & T)[];
